Added a health bar and game over screens. Feel free to mess with these.

+++ app/src/main/java/com/example/hukin/Logic/PlayerStatus.java
@@ -32,7 +32,7 @@ public class PlayerStatus extends Moveable {
     }
 
     // A whole bunch of self-explanatory get functions
-    public double getPlayerRole() {
+    public int getPlayerRole() {
         return playerRole;
     }
